refactor: :recycle: added endpoint /user/register
