+++ src/routes/v1.ts
@@ -1,9 +1,9 @@
 import { Request, Response, Router } from "express";
 import EndpointRoot from "../endpoints/root";
-import EndpointUser from "../endpoints/user";
 import { EndpointUserData } from "../endpoints/user/data";
 import { EndpointUserDelete } from "../endpoints/user/delete";
 import { EndpointUserLogin } from "../endpoints/user/login";
+import { EndpointUserRegister } from "../endpoints/user/register";
 import { EndpointUserRenew } from "../endpoints/user/renew";
 import EndpointUserAdmin from "../endpoints/user_admin";
 import Routes from "../interfaces/routes";
@@ -29,9 +29,7 @@ export class V1Routes extends Routes {
         this.router.all('/', (request: Request, response: Response) => {
             new EndpointRoot(request, response).method();
         });
-        this.router.all('/user', (request: Request, response: Response) => {
-            new EndpointUser(request, response).method();
-        });
+
         this.router.all('/user/admin', (request: Request, response: Response) => {
             new EndpointUserAdmin(request, response).method();
         });
@@ -45,6 +43,9 @@ export class V1Routes extends Routes {
         });
         this.router.all('/user/login', (request: Request, response: Response) => {
             new EndpointUserLogin(request, response).method();
+        });
+        this.router.all('/user/register', (request: Request, response: Response) => {
+            new EndpointUserRegister(request, response).method();
         });
         this.router.all('/user/renew', (request: Request, response: Response) => {
             new EndpointUserRenew(request, response).method();
